1.1.2 - Count now shows to decimal

+++ Diadochokinetic Assess/Models/Models.swift
@@ -13,7 +13,7 @@ struct Record : Identifiable, Hashable {
     var date: Date
     var taps: Int
     var timed: Bool
-    var duration: Int
+    var duration: Double
 }
 
 enum PossibleCountingStates {
